+++ panda/board/safety/safety_chrysler.h
@@ -8,6 +8,16 @@ const SteeringLimits CHRYSLER_STEERING_LIMITS = {
   .type = TorqueMotorLimited,
 };
 
+const SteeringLimits CHRYSLER_JEEPS_STEERING_LIMITS = {
+  .max_steer = 261,
+  .max_rt_delta = 225, // 6 max rate up * 100Hz send rate * 250000 RT interval / 1000000 = 150 ; 150 * 1.5 for safety pad = 225
+  .max_rt_interval = 250000,
+  .max_rate_up = 6,
+  .max_rate_down = 6,
+  .max_torque_error = 160,
+  .type = TorqueMotorLimited,
+};
+
 const SteeringLimits CHRYSLER_RAM_DT_STEERING_LIMITS = {
   .max_steer = 350,
   .max_rt_delta = 112,
@@ -34,7 +44,9 @@ typedef struct {
   const int ESP_8;
   const int ECM_5;
   const int DAS_3;
+  const int DAS_5;
   const int DAS_6;
+  const int GEAR;
   const int LKAS_COMMAND;
   const int LKAS_HEARTBIT;
   const int CRUISE_BUTTONS;
@@ -46,8 +58,10 @@ const ChryslerAddrs CHRYSLER_ADDRS = {
   .ESP_1            = 0x140,  // Brake pedal and vehicle speed
   .ESP_8            = 0x11C,  // Brake pedal and vehicle speed
   .ECM_5            = 0x22F,  // Throttle position sensor
-  .DAS_3            = 0x1F4,  // ACC engagement states from DASM
+  .DAS_3            = 0x1F4,  // ACC
+  .DAS_5            = 0x271,  // ACC for hybrids
   .DAS_6            = 0x2A6,  // LKAS HUD and auto headlight control from DASM
+  .GEAR             = 0x170,  // Current GEAR
   .LKAS_COMMAND     = 0x292,  // LKAS controls from DASM
   .LKAS_HEARTBIT    = 0x2D9,  // LKAS HEARTBIT from DASM
   .CRUISE_BUTTONS   = 0x23B,  // Cruise control buttons
@@ -81,6 +95,8 @@ const CanMsg CHRYSLER_TX_MSGS[] = {
   {CHRYSLER_ADDRS.CRUISE_BUTTONS, 0, 3},
   {CHRYSLER_ADDRS.LKAS_COMMAND, 0, 6},
   {CHRYSLER_ADDRS.DAS_6, 0, 8},
+  {CHRYSLER_ADDRS.DAS_3, 0, 8},
+  {CHRYSLER_ADDRS.DAS_5, 0, 8},
   {CHRYSLER_ADDRS.LKAS_HEARTBIT, 0, 5},
 };
 
@@ -103,6 +119,7 @@ RxCheck chrysler_rx_checks[] = {
   {.msg = {{514, 0, 8, .check_checksum = false, .max_counter = 0U, .frequency = 100U}, { 0 }, { 0 }}},
   {.msg = {{CHRYSLER_ADDRS.ECM_5, 0, 8, .check_checksum = true, .max_counter = 15U, .frequency = 50U}, { 0 }, { 0 }}},
   {.msg = {{CHRYSLER_ADDRS.DAS_3, 0, 8, .check_checksum = true, .max_counter = 15U, .frequency = 50U}, { 0 }, { 0 }}},
+  {.msg = {{CHRYSLER_ADDRS.GEAR, 0, 8, .check_checksum = true, .max_counter = 15U, .frequency = 50U}, { 0 }, { 0 }}},
 };
 
 RxCheck chrysler_ram_dt_rx_checks[] = {
@@ -125,13 +142,17 @@ RxCheck chrysler_ram_hd_rx_checks[] = {
 
 const uint32_t CHRYSLER_PARAM_RAM_DT = 1U;  // set for Ram DT platform
 const uint32_t CHRYSLER_PARAM_RAM_HD = 2U;  // set for Ram HD platform
+const uint32_t CHRYSLER_PARAM_JEEP = 4U;  // set for Jeep platform
 
 typedef enum {
   CHRYSLER_RAM_DT,
   CHRYSLER_RAM_HD,
-  CHRYSLER_PACIFICA,  // plus Jeep
+  CHRYSLER_PACIFICA,
+  CHRYSLER_JEEP,
 } ChryslerPlatform;
-ChryslerPlatform chrysler_platform = CHRYSLER_PACIFICA;
+ChryslerPlatform chrysler_platform = CHRYSLER_JEEP;
+bool ram_platform = false;
+
 const ChryslerAddrs *chrysler_addrs = &CHRYSLER_ADDRS;
 
 static uint32_t chrysler_get_checksum(const CANPacket_t *to_push) {
@@ -186,35 +207,49 @@ static void chrysler_rx_hook(const CANPacket_t *to_push) {
     update_sample(&torque_meas, torque_meas_new);
   }
 
-  // enter controls on rising edge of ACC, exit controls on ACC off
-  const int das_3_bus = (chrysler_platform == CHRYSLER_PACIFICA) ? 0 : 2;
+  if ((bus == 0) && (addr == chrysler_addrs->GEAR)) {
+    forward_gear = ((GET_BYTE(to_push, 0) >> 2) & 0x7U) >= 4;
+  }
+
+  const int das_3_bus = (ram_platform) ? 2 : 0;
   if ((bus == das_3_bus) && (addr == chrysler_addrs->DAS_3)) {
-    bool cruise_engaged = GET_BIT(to_push, 21U) || ((alternative_experience & ALT_EXP_AOLC_ENABLED) && GET_BIT(to_push, 20U));
-    pcm_cruise_check(cruise_engaged);
+    if (forward_gear) {
+      const bool cruise_available = GET_BIT(to_push, 20U);
+      const bool lkas_enabled = GET_BIT(to_push, 21U) || ((alternative_experience & ALT_EXP_AOLC_ENABLED) && cruise_available);
+      pcm_cruise_check(lkas_enabled);
+
+      long_allowed = !cruise_available && (alternative_experience & ALT_EXP_LONG_ENABLED);
+      if (long_allowed) {
+        pcm_cruise_check(false);
+        pcm_cruise_check(true);
+      }
+    } else {
+      pcm_cruise_check(false);
+      long_allowed = false;
+    }
   }
 
   // TODO: use the same message for both
   // update vehicle moving
-  if ((chrysler_platform != CHRYSLER_PACIFICA) && (bus == 0) && (addr == chrysler_addrs->ESP_8)) {
+  if ((ram_platform) && (bus == 0) && (addr == chrysler_addrs->ESP_8)) {
     vehicle_moving = ((GET_BYTE(to_push, 4) << 8) + GET_BYTE(to_push, 5)) != 0U;
-  }
-  if ((chrysler_platform == CHRYSLER_PACIFICA) && (bus == 0) && (addr == 514)) {
+  } else if ((bus == 0) && (addr == 514)) {
     int speed_l = (GET_BYTE(to_push, 0) << 4) + (GET_BYTE(to_push, 1) >> 4);
     int speed_r = (GET_BYTE(to_push, 2) << 4) + (GET_BYTE(to_push, 3) >> 4);
     vehicle_moving = (speed_l != 0) || (speed_r != 0);
   }
 
-  // exit controls on rising edge of gas press
   if ((bus == 0) && (addr == chrysler_addrs->ECM_5)) {
     gas_pressed = GET_BYTE(to_push, 0U) != 0U;
   }
 
-  // exit controls on rising edge of brake press
   if ((bus == 0) && (addr == chrysler_addrs->ESP_1)) {
     brake_pressed = ((GET_BYTE(to_push, 0U) & 0xFU) >> 2U) == 1U;
   }
 
-  generic_rx_checks((bus == 0) && (addr == chrysler_addrs->LKAS_COMMAND));
+  if (!long_allowed) {
+    generic_rx_checks((bus == 0) && (addr == chrysler_addrs->LKAS_COMMAND));
+  }
 }
 
 static bool chrysler_tx_hook(const CANPacket_t *to_send) {
@@ -223,17 +258,24 @@ static bool chrysler_tx_hook(const CANPacket_t *to_send) {
 
   // STEERING
   if (addr == chrysler_addrs->LKAS_COMMAND) {
-    int start_byte = (chrysler_platform == CHRYSLER_PACIFICA) ? 0 : 1;
+    int start_byte = (ram_platform) ? 1 : 0;
     int desired_torque = ((GET_BYTE(to_send, start_byte) & 0x7U) << 8) | GET_BYTE(to_send, start_byte + 1);
     desired_torque -= 1024;
 
-    const SteeringLimits limits = (chrysler_platform == CHRYSLER_PACIFICA) ? CHRYSLER_STEERING_LIMITS :
-                                  (chrysler_platform == CHRYSLER_RAM_DT) ? CHRYSLER_RAM_DT_STEERING_LIMITS : CHRYSLER_RAM_HD_STEERING_LIMITS;
+    const SteeringLimits limits = chrysler_platform == CHRYSLER_JEEP ? CHRYSLER_JEEPS_STEERING_LIMITS :
+                                  chrysler_platform == CHRYSLER_RAM_DT ? CHRYSLER_RAM_DT_STEERING_LIMITS :
+                                  chrysler_platform == CHRYSLER_RAM_HD ? CHRYSLER_RAM_HD_STEERING_LIMITS : CHRYSLER_STEERING_LIMITS;
 
-    bool steer_req = (chrysler_platform == CHRYSLER_PACIFICA) ? GET_BIT(to_send, 4U) : (GET_BYTE(to_send, 3) & 0x7U) == 2U;
+    bool steer_req = (ram_platform) ? (GET_BYTE(to_send, 3) & 0x7U) == 2U : GET_BIT(to_send, 4U);
+
     if (steer_torque_cmd_checks(desired_torque, steer_req, limits)) {
       tx = false;
     }
+  }
+
+  // block long from sending ACC when a pedal is pressed
+  if (addr == chrysler_addrs->DAS_3 || addr == chrysler_addrs->DAS_5) {
+    tx = !brake_pressed && !gas_pressed;
   }
 
 //  // FORCE CANCEL: only the cancel button press is allowed
@@ -271,17 +313,23 @@ static safety_config chrysler_init(uint16_t param) {
 
   bool enable_ram_dt = GET_FLAG(param, CHRYSLER_PARAM_RAM_DT);
   if (enable_ram_dt) {
+    ram_platform = true;
     chrysler_platform = CHRYSLER_RAM_DT;
     chrysler_addrs = &CHRYSLER_RAM_DT_ADDRS;
     ret = BUILD_SAFETY_CFG(chrysler_ram_dt_rx_checks, CHRYSLER_RAM_DT_TX_MSGS);
 #ifdef ALLOW_DEBUG
   } else if (GET_FLAG(param, CHRYSLER_PARAM_RAM_HD)) {
+    ram_platform = true;
     chrysler_platform = CHRYSLER_RAM_HD;
     chrysler_addrs = &CHRYSLER_RAM_HD_ADDRS;
     ret = BUILD_SAFETY_CFG(chrysler_ram_hd_rx_checks, CHRYSLER_RAM_HD_TX_MSGS);
 #endif
   } else {
-    chrysler_platform = CHRYSLER_PACIFICA;
+    if (GET_FLAG(param, CHRYSLER_PARAM_JEEP)) {
+      chrysler_platform = CHRYSLER_JEEP;
+    } else {
+      chrysler_platform = CHRYSLER_PACIFICA;
+    }
     chrysler_addrs = &CHRYSLER_ADDRS;
     ret = BUILD_SAFETY_CFG(chrysler_rx_checks, CHRYSLER_TX_MSGS);
   }